wine_detection_api/requirements.txt
flask==2.0.1
ultralytics
opencv-python
numpy
python-dotenv
